chore: update landingpage
