+++ src/page-building/landingpage/blog.tsx
@@ -1,0 +1,34 @@
+import { useTranslation } from 'gatsby-plugin-react-i18next';
+import { TeaserGrid } from '../../components/teasers/teaser-grid';
+import { Teaser } from '../../components/teasers/teaser';
+import React from 'react';
+import { SyPersonioJob } from '../../@types/personio';
+import { HomePageHeaderBlock } from './support';
+
+interface CareerProps {
+  positions: SyPersonioJob[];
+}
+
+export const Blog = ({ posts }: CareerProps) => {
+  const { t } = useTranslation();
+
+  return (
+    <>
+      <HomePageHeaderBlock
+        topline={t('main.career.topline')}
+        headline={t('main.career.title')}
+        large={true}
+      >
+        {t('main.career.text')}
+      </HomePageHeaderBlock>
+
+      <TeaserGrid>
+        {posts.map((item) => (
+          <Teaser key={item.id} title={item.name} linkTo={item.fields?.path}>
+            {item.short}
+          </Teaser>
+        ))}
+      </TeaserGrid>
+    </>
+  );
+};
