+++ src/page-building/landingpage/support.tsx
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+import { HeaderBlock } from '../../components/header-block/header-block';
+
+export const HomePageHeaderBlock = styled(HeaderBlock)`
+  margin-bottom: 80px;
+`;
