+++ src/page-building/service/products-services.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { ContentBlockContainer } from '../../components/layout/content-block-container';
 import { useTranslation } from 'gatsby-plugin-react-i18next';
-import { ExpandableStyled, Intro, UnorderedList } from './support';
+import { Intro, UnorderedList } from './support';
 
 export const ProductsServices = () => {
   const { t } = useTranslation();
